Update command box to mask password properly
