+++ src/main/java/seedu/investigapptor/ui/CommandBox.java
@@ -1,12 +1,10 @@
 package seedu.investigapptor.ui;
 
-import static seedu.investigapptor.logic.parser.CliSyntax.PREFIX_PASSWORD;
-
 import java.util.logging.Logger;
 
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.scene.control.TextField;
+import javafx.scene.control.PasswordField;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.Region;
 import seedu.investigapptor.commons.core.LogsCenter;
@@ -16,6 +14,7 @@ import seedu.investigapptor.logic.Logic;
 import seedu.investigapptor.logic.commands.CommandResult;
 import seedu.investigapptor.logic.commands.exceptions.CommandException;
 import seedu.investigapptor.logic.parser.exceptions.ParseException;
+import seedu.investigapptor.ui.skin.PasswordFieldSkin;
 
 /**
  * The UI component that is responsible for receiving user command inputs.
@@ -31,23 +30,14 @@ public class CommandBox extends UiPart<Region> {
     private boolean hideEnabled;
 
     @FXML
-    private TextField commandTextField;
-
-    @FXML
-    private TextField commandTextDisplay;
+    private PasswordField commandTextField;
 
     public CommandBox(Logic logic) {
         super(FXML);
         this.logic = logic;
         // calls #setStyleToDefault() whenever there is a change to the text of the command box.
-        commandTextDisplay.textProperty().addListener((unused1, unused2, unused3) -> setStyleToDefault());
-        commandTextField.setOpacity(0);
-
-        // checks for password prefix
-        commandTextField.textProperty()
-                .addListener((observable, oldValue, newValue) -> commandTextDisplay
-                        .setText(hidePasswordText(newValue)));
-        //commandTextField.setSkin(new TextFieldCaret(commandTextField));
+        commandTextField.textProperty().addListener((unused1, unused2, unused3) -> setStyleToDefault());
+        commandTextField.setSkin(new PasswordFieldSkin(commandTextField));
 
         hideEnabled = true;
         historySnapshot = logic.getHistorySnapshot();
@@ -72,10 +62,6 @@ public class CommandBox extends UiPart<Region> {
         case ESCAPE:
             keyEvent.consume();
             clearScreenText();
-            break;
-        case CONTROL:
-            keyEvent.consume();
-            togglePasswordHide();
             break;
         default:
                 // let JavaFx handle the keypress
@@ -147,49 +133,10 @@ public class CommandBox extends UiPart<Region> {
      */
     private void clearScreenText() {
         commandTextField.setText("");
-        commandTextDisplay.setText("");
     }
 
-    /**
-     *  Toggles between hiding the password and revealing the password field to the user
-     */
-    private void togglePasswordHide() {
-        if (hideEnabled) {
-            commandTextField.setOpacity(1);
-            commandTextDisplay.setOpacity(0);
-            hideEnabled = false;
-        } else {
-            commandTextField.setOpacity(0);
-            commandTextDisplay.setOpacity(1);
-            hideEnabled = true;
-        }
-    }
-    /**
-     * Hides password string
-     * @param inputText
-     * @return
-     */
-    private String hidePasswordText(String inputText) {
-        StringBuilder sb = new StringBuilder(inputText);
-        int prefixIndex = inputText.indexOf(PREFIX_PASSWORD.getPrefix());
 
-        if (hasPasswordPrefix(inputText)) {
-            for (int i = prefixIndex + 3; i < inputText.length(); i++) {
-                sb.setCharAt(i, '*');
-            }
-        }
-        return sb.toString();
-    }
 
-    /**
-     * Checks for presence of password prefix
-     * @param inputText
-     * @return
-     */
-    private boolean hasPasswordPrefix(String inputText) {
-        int passwordPrefixIndex = inputText.indexOf(PREFIX_PASSWORD.getPrefix());
-        return passwordPrefixIndex != -1;
-    }
     //@@author
 
     /**
@@ -206,7 +153,6 @@ public class CommandBox extends UiPart<Region> {
      * Sets the command box style to use the default style.
      */
     private void setStyleToDefault() {
-        commandTextDisplay.getStyleClass().remove(ERROR_STYLE_CLASS);
         commandTextField.getStyleClass().remove(ERROR_STYLE_CLASS);
     }
 
@@ -214,14 +160,7 @@ public class CommandBox extends UiPart<Region> {
      * Sets the command box style to indicate a failed command.
      */
     private void setStyleToIndicateCommandFailure() {
-        ObservableList<String> styleClassDisplay = commandTextDisplay.getStyleClass();
         ObservableList<String> styleClassField = commandTextField.getStyleClass();
-
-        if (styleClassDisplay.contains(ERROR_STYLE_CLASS)) {
-            return;
-        }
-
-        styleClassDisplay.add(ERROR_STYLE_CLASS);
 
         if (styleClassField.contains(ERROR_STYLE_CLASS)) {
             return;
